refactor: move UIColorPaletteManager to titanicsend.ui.color
